Move JobManager out of .core.utils into .utils
This will allow non-core tasks to have access to the JobManager class
without requiring the core tasks/plugins/etc.
