+++ tests/io_test.py
@@ -1,11 +1,14 @@
 import copy
 import httmock
 import os
-import girder_worker
 import shutil
 import unittest
 
-from girder_worker.core.utils import JobStatus
+import girder_worker
+from girder_worker.utils import JobStatus
+
+import girder_worker.core
+
 
 _tmp = None
 
@@ -148,7 +151,7 @@ with open(file) as f:
             }
         }
 
-        job_mgr = girder_worker.core.utils.JobManager(
+        job_mgr = girder_worker.utils.JobManager(
             True, url='http://jobstatus/')
 
         received = []
@@ -231,7 +234,7 @@ with open(file) as f:
         self.assertRegexpMatches(outputs['_tempdir']['data'], _tmp + '.+')
 
     def testConvertingStatus(self):
-        job_mgr = girder_worker.core.utils.JobManager(
+        job_mgr = girder_worker.utils.JobManager(
             True, url='http://jobstatus/')
 
         status_changes = []
